pyproject.toml
[tool.poetry]
name = "subgraph_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
psycopg2-binary = "^2.8.6"
python-telegram-bot = "^13.1"
sqlalchemy = "^1.3.20"
sqlalchemy-utils = "^0.36.8"
requests = "^2.25.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
